Add basic required field validation and logic for imports.
Required fields are passed in a comma-delimited format via a "data-required" attribute in the form tag. When the "Process Import" button is clicked, those fields (if set) are simply checked to see if a field has been selected for mapping. If not, form submission is skipped and a class is added to the parent table row indicating that the field is "required".

TODO: Localized validation message injected into the "Data Preview" column for the field.